Return unsupported_response_type error for invalid response types
This is a OIDC conformance reimplementation of exception handling
in which a proper error message is returned to OIDC client
if response type is invalid (error unsupported_response_type is
returned to the OIDC client using HTTP redirect response).
This could be replaced by simply showing an error page (according
to the conformance test), so this could be done much simpler
(less need for overriding league/oauth2 package).